feat(diff): Allow unlimited diff size and fix prompt
Change the default diff character limit to 0 to represent an unlimited
size. The script now only truncates the diff if the limit is explicitly
set to a value greater than 0.

This allows the full diff to be sent to the model by default, which can
improve the quality of the generated commit message.

Additionally, this fixes a bug in the prompt template that caused the
diff content to be improperly wrapped, potentially corrupting the
output.